requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
astunparse==1.6.3
attrs==26.1.0
bce-python-sdk==0.9.64
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
boto3==1.34.69
botocore==1.34.162
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
chardet==7.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
crispy-bootstrap5==0.7
cryptography==46.0.6
cycler==0.12.1
datasets==4.8.4
deepface==0.0.99
defusedxml==0.7.1
dill==0.4.1
dj-database-url==3.1.2
Django==4.2.9
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.7.0
django-phonenumber-field==8.4.0
django-redis==5.4.0
django-storages==1.14.2
django-two-factor-auth==1.18.1
django-unfold==0.89.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
easyocr==1.7.2
face_recognition_models==0.3.0
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==21.2.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
humanize==4.9.0
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
keras==3.13.2
kiwisolver==1.5.0
kombu==5.6.2
lazy-loader==0.5
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
modelscope==1.35.1
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.7.1
multiprocess==0.70.19
mysql-connector-python==9.6.0
mysqlclient==2.2.1
namex==0.1.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
oauthlib==3.3.1
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt-einsum==3.3.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pillow==10.2.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
pypdfium2==5.6.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2026.1.post1
PyYAML==6.0.2
qrcode==8.2
redis==5.0.3
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.3.3
ruamel.yaml==0.19.1
s3transfer==0.10.4
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
text-unidecode==1.3
tf_keras==2.21.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.6
whitenoise==6.6.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
